# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Echobot",
    "description": "AI驱动的数字人动画平台 API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/digital-human/preview": {
      "post": {
        "tags": [
          "数字人"
        ],
        "summary": "创建数字人预览任务",
        "description": "创建数字人预览任务\n\n上传家长的正面照片，系统会使用阿里云 EMO 技术生成动态数字人预览视频。\n这是一个异步任务，创建后需要轮询 GET /preview/{task_id} 获取结果。\n\n**图片要求:**\n- 格式：JPG 或 PNG\n- 内容：清晰的正面人像照片\n- 大小：最大10MB\n\n**音频来源（三选一）:**\n1. **不传音频参数** - 使用系统默认音色生成预览\n2. **上传 audio 文件** - 直接使用上传的音频驱动数字人\n3. **传 voice_profile_id + preview_text** - 使用已保存的克隆声音合成音频\n\n**请求参数:**\n- **image**: 头像照片（multipart/form-data，必填）\n- **audio**: 音频文件（可选，WAV/MP3格式，最大20MB）\n- **voice_profile_id**: 声音档案ID（可选）\n- **preview_text**: 预览文本（使用声音档案时必填）\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": {\n        \"task_id\": \"user123_1703980800000\",\n        \"status\": \"processing\",\n        \"audio_source\": \"default\",\n        \"message\": \"Digital human task created, poll for status\"\n    }\n}\n```",
        "operationId": "create_digital_human_preview_api_v1_digital_human_preview_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_digital_human_preview_api_v1_digital_human_preview_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/digital-human/preview/{task_id}": {
      "get": {
        "tags": [
          "数字人"
        ],
        "summary": "获取数字人任务状态",
        "description": "获取数字人任务状态\n\n轮询此接口获取任务的处理进度和结果。\n建议轮询间隔：3-5秒（数字人生成较慢）。\n\n**路径参数:**\n- **task_id**: 任务ID（从 POST /preview 返回）\n\n**任务状态说明:**\n- **detecting**: 图像检测中，检测人脸和关键点\n- **generating**: 视频生成中，EMO 模型处理\n- **completed**: 完成，video_url 包含生成的视频\n- **failed**: 失败，error 包含错误信息\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": {\n        \"task_id\": \"user123_1703980800000\",\n        \"status\": \"completed\",\n        \"progress\": 100,\n        \"video_url\": \"/uploads/avatar/user123/preview.mp4\",\n        \"error\": null\n    }\n}\n```",
        "operationId": "get_digital_human_status_api_v1_digital_human_preview__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/digital-human/profiles": {
      "get": {
        "tags": [
          "数字人"
        ],
        "summary": "获取头像档案列表",
        "description": "获取用户的所有头像档案\n\n返回当前用户保存的所有头像档案列表。\n这些头像可用于故事生成功能。\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": {\n        \"profiles\": [\n            {\n                \"id\": \"694534f1c32ffbd6151c18a4\",\n                \"name\": \"爸爸的头像\",\n                \"image_url\": \"/uploads/avatar/user123/photo.jpg\",\n                \"preview_video_url\": \"/uploads/avatar/user123/preview.mp4\",\n                \"created_at\": \"2025-01-01T12:00:00\"\n            }\n        ],\n        \"total\": 2\n    }\n}\n```",
        "operationId": "get_avatar_profiles_api_v1_digital_human_profiles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "数字人"
        ],
        "summary": "保存头像档案",
        "description": "保存头像档案\n\n数字人预览完成且满意后，将头像保存到头像档案库。\n保存后可在故事生成功能中重复使用该头像。\n\n**请求参数:**\n- **task_id**: 已完成的数字人任务ID\n- **name**: 头像名称，如：爸爸的头像、妈妈的头像\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": {\n        \"profile\": {\n            \"id\": \"694534f1c32ffbd6151c18a4\",\n            \"name\": \"爸爸的头像\",\n            \"image_url\": \"/uploads/avatar/user123/photo.jpg\",\n            \"preview_video_url\": \"/uploads/avatar/user123/preview.mp4\",\n            \"created_at\": \"2025-01-01T12:00:00\"\n        }\n    }\n}\n```",
        "operationId": "save_avatar_profile_api_v1_digital_human_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveAvatarProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/digital-human/profiles/{profile_id}": {
      "get": {
        "tags": [
          "数字人"
        ],
        "summary": "获取头像档案详情",
        "description": "获取单个头像档案详情\n\n**路径参数:**\n- **profile_id**: 头像档案ID\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": {\n        \"id\": \"694534f1c32ffbd6151c18a4\",\n        \"name\": \"爸爸的头像\",\n        \"image_url\": \"/uploads/avatar/user123/photo.jpg\",\n        \"preview_video_url\": \"/uploads/avatar/user123/preview.mp4\",\n        \"created_at\": \"2025-01-01T12:00:00\"\n    }\n}\n```",
        "operationId": "get_avatar_profile_api_v1_digital_human_profiles__profile_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "数字人"
        ],
        "summary": "更新头像档案",
        "description": "更新头像档案名称\n\n**路径参数:**\n- **profile_id**: 头像档案ID\n\n**请求参数:**\n- **name**: 新的头像名称\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"更新成功\",\n    \"data\": null\n}\n```",
        "operationId": "update_avatar_profile_api_v1_digital_human_profiles__profile_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAvatarProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "数字人"
        ],
        "summary": "删除头像档案",
        "description": "删除头像档案\n\n删除后将无法恢复，且使用该头像的历史记录不受影响。\n\n**路径参数:**\n- **profile_id**: 头像档案ID\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"删除成功\",\n    \"data\": null\n}\n```",
        "operationId": "delete_avatar_profile_api_v1_digital_human_profiles__profile_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_digital_human_preview_api_v1_digital_human_preview_post": {
        "properties": {
          "image": {
            "type": "string",
            "format": "binary",
            "title": "Image",
            "description": "头像照片，清晰正面人像，JPG/PNG格式"
          },
          "audio": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio",
            "description": "音频文件（可选），用于数字人说话"
          },
          "voice_profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Profile Id",
            "description": "已保存的声音档案ID（可选）"
          },
          "preview_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Text",
            "description": "预览文本，使用声音档案时需要提供"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_create_digital_human_preview_api_v1_digital_human_preview_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SaveAvatarProfileRequest": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "description": "已完成的数字人预览任务ID",
            "example": "user123_1703980800000"
          },
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Name",
            "description": "头像名称，便于识别，如：爸爸的头像、妈妈的头像",
            "example": "爸爸的头像"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "name"
        ],
        "title": "SaveAvatarProfileRequest",
        "description": "保存头像档案请求\n\n数字人预览成功后，将头像保存为档案。\n保存后可在故事生成功能中重复使用。",
        "example": {
          "name": "爸爸的头像",
          "task_id": "user123_1703980800000"
        }
      },
...
      "UpdateAvatarProfileRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Name",
            "description": "新的头像名称",
            "example": "妈妈的头像"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "UpdateAvatarProfileRequest",
        "description": "更新头像档案请求\n\n用于修改头像档案的名称。",
        "example": {
          "name": "妈妈的头像"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}